WI #442 Move all performance test to Performance class.
Add "-Incremental" in the name of the report of Incremental.
Remove decimal part of average time to ease the reading.
