+++ TypeCobol.Test/Parser/Incremental/TestIncrementalPipeline.cs
@@ -12,95 +12,6 @@ using TypeCobol.Test.Utils;
 
 namespace TypeCobol.Test.Parser.Incremental {
 
-	[TestClass]
-	public class TestIncrementalPipeline {
-
-		[TestMethod]
-		[TestCategory("Incremental")]
-		[TestProperty("Time","long")]
-		public void CheckPerformance() {
-            // Sample program properties
-            string folder = "Parser" + Path.DirectorySeparatorChar + "Samples";
-            string textName = "BigBatch";
-            DocumentFormat documentFormat = DocumentFormat.RDZReferenceFormat;
-
-            // Create a FileCompiler for this program
-            DirectoryInfo localDirectory = new DirectoryInfo(PlatformUtils.GetPathForProjectFile(folder));
-            if (!localDirectory.Exists)
-            {
-                throw new Exception(String.Format("Directory : {0} does not exist", localDirectory.FullName));
-            }
-            CompilationProject project = new CompilationProject("test",
-                localDirectory.FullName, new string[] { ".cbl", ".cpy" },
-                documentFormat.Encoding, documentFormat.EndOfLineDelimiter, documentFormat.FixedLineLength, documentFormat.ColumnsLayout, new TypeCobolOptions());
-            FileCompiler compiler = new FileCompiler(null, textName, project.SourceFileProvider, project, documentFormat.ColumnsLayout, new TypeCobolOptions(), null, false, project);
-            //Make an incremental change to the source code
-            TestUtils.CompilationStats stats = new TestUtils.CompilationStats();
-            ExecuteInceremental(compiler, stats);
-
-            // Display a performance report
-            StringBuilder report = new StringBuilder();
-            report.AppendLine("Program properties :");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.CobolTextLines.Count + " lines");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.CodeElementsDocumentSnapshot.CodeElements.Count() + " code elements");
-            report.AppendLine("First compilation performance");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.PerfStatsForText.FirstCompilationTime + " ms : text update");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.PerfStatsForScanner.FirstCompilationTime + " ms : scanner");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.PerfStatsForPreprocessor.FirstCompilationTime + " ms : preprocessor");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.PerfStatsForCodeElementsParser.FirstCompilationTime + " ms : code elements parser");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.PerfStatsForTemporarySemantic.FirstCompilationTime + " ms : temporary semantic class parser");
-            report.AppendLine("- " + compiler.CompilationResultsForProgram.PerfStatsForProgramCrossCheck.FirstCompilationTime + " ms : cross check class parser");
-            report.AppendLine("Incremental compilation performance");
-            report.AppendLine("- " + stats.AverageTextUpdateTime + " ms : text update");
-            report.AppendLine("- " + stats.AverageScannerTime + " ms : scanner");
-            report.AppendLine("- " + stats.AveragePreprocessorTime + " ms : preprocessor");
-            report.AppendLine("- " + stats.AverageCodeElementParserTime + " ms : code elements parser");
-            report.AppendLine("- " + stats.AverateTemporarySemanticsParserTime + " ms : temporary semantic class parser");
-            report.AppendLine("- " + stats.AverageCrossCheckerParserTime + " ms : cross check class parser");
-            TestUtils.CreateRunReport(TestUtils.GetReportDirectoryPath(), compiler.CobolFile.Name, compiler.CompilationResultsForProgram, stats);
-        }
-
-        public static void ExecuteInceremental(FileCompiler compiler,TestUtils.CompilationStats stats)
-        {
-            // Execute a first (complete) compilation
-            compiler.CompileOnce();
-            //Iterate multiple times over an incremental change
-            int incrementalIterationNumber = 20;
-            for (int i = 0; i < incrementalIterationNumber; i++)
-            {
-                // Append one line in the middle of the program
-                ITextLine newLine = new TextLineSnapshot(9211, "094215D    DISPLAY '-ICLAUA      = ' ICLAUA.                            0000000", null);
-                TextChangedEvent textChangedEvent = new TextChangedEvent();
-                textChangedEvent.TextChanges.Add(new TextChange(TextChangeType.LineInserted, 9211, newLine));
-                compiler.CompilationResultsForProgram.UpdateTextLines(textChangedEvent);
-
-                // Execute a second (incremental) compilation
-                compiler.CompileOnce();
-                //Accumulate results
-                stats.AverageTextUpdateTime += compiler.CompilationResultsForProgram.PerfStatsForText.LastRefreshTime;
-                stats.AverageScannerTime += compiler.CompilationResultsForProgram.PerfStatsForScanner.LastRefreshTime;
-                stats.AveragePreprocessorTime +=
-                    compiler.CompilationResultsForProgram.PerfStatsForPreprocessor.LastRefreshTime;
-                stats.AverageCodeElementParserTime += compiler.CompilationResultsForProgram
-                    .PerfStatsForCodeElementsParser.LastRefreshTime;
-                stats.AverateTemporarySemanticsParserTime += compiler.CompilationResultsForProgram
-                    .PerfStatsForTemporarySemantic.LastRefreshTime;
-                stats.AverageCrossCheckerParserTime += compiler.CompilationResultsForProgram
-                    .PerfStatsForProgramCrossCheck.LastRefreshTime;
-            }
-            //Compute average time needed for each phase
-            stats.AverageTextUpdateTime = stats.AverageTextUpdateTime / (float)incrementalIterationNumber;
-            stats.AverageScannerTime = stats.AverageScannerTime / (float)incrementalIterationNumber;
-            stats.AveragePreprocessorTime = stats.AveragePreprocessorTime / (float) incrementalIterationNumber;
-            stats.AverageCodeElementParserTime =
-                stats.AverageCodeElementParserTime / (float) incrementalIterationNumber;
-            stats.AverateTemporarySemanticsParserTime =
-                stats.AverateTemporarySemanticsParserTime / (float) incrementalIterationNumber;
-            stats.AverageCrossCheckerParserTime =
-                stats.AverageCrossCheckerParserTime / (float) incrementalIterationNumber;
-
-        }
-    }
 
     [TestClass]
     public class TestChangeEvent
